update profile UI
I have make the UI more attractive
