+++ lib/util/profilegrid.dart
@@ -1,262 +1,263 @@
-import 'package:flutter/material.dart';
-import 'package:taskmatch/util/update_profile_form.dart';
+// import 'package:flutter/material.dart';
 
-void main() {
-  runApp(const MyApp());
-}
+// import '../pages/profilepage/update_profile_page.dart';
 
-class MyApp extends StatelessWidget {
-  const MyApp({Key? key}) : super(key: key);
+// void main() {
+//   runApp(const MyApp());
+// }
 
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      theme: ThemeData(
-        primaryColor: Colors.green,
-        textTheme: const TextTheme(
-          bodyLarge: TextStyle(color: Colors.white),
-          bodyMedium: TextStyle(color: Colors.white),
-          headlineMedium: TextStyle(color: Colors.white),
-          headlineSmall: TextStyle(color: Colors.white),
-          titleLarge: TextStyle(color: Colors.white),
-        ),
-      ),
-      home: const RootPage(),
-    );
-  }
-}
+// class MyApp extends StatelessWidget {
+//   const MyApp({Key? key}) : super(key: key);
 
-class RootPage extends StatefulWidget {
-  const RootPage({Key? key}) : super(key: key);
+//   @override
+//   Widget build(BuildContext context) {
+//     return MaterialApp(
+//       debugShowCheckedModeBanner: false,
+//       theme: ThemeData(
+//         primaryColor: Colors.green,
+//         textTheme: const TextTheme(
+//           bodyLarge: TextStyle(color: Colors.white),
+//           bodyMedium: TextStyle(color: Colors.white),
+//           headlineMedium: TextStyle(color: Colors.white),
+//           headlineSmall: TextStyle(color: Colors.white),
+//           titleLarge: TextStyle(color: Colors.white),
+//         ),
+//       ),
+//       home: const RootPage(),
+//     );
+//   }
+// }
 
-  @override
-  State<RootPage> createState() => _RootPageState();
-}
+// class RootPage extends StatefulWidget {
+//   const RootPage({Key? key}) : super(key: key);
 
-class _RootPageState extends State<RootPage> {
-  int currentPage = 0;
-  int _selectedIndex = 2;
-  int currentIndex = 0;
+//   @override
+//   State<RootPage> createState() => _RootPageState();
+// }
 
-  void _onItemTapped(int index) {
-    setState(() {
-      _selectedIndex = index;
-    });
+// class _RootPageState extends State<RootPage> {
+//   int currentPage = 0;
+//   int _selectedIndex = 2;
+//   int currentIndex = 0;
 
-    if (_selectedIndex == 1) {
-      _buildSearchBar();
-    }
+//   void _onItemTapped(int index) {
+//     setState(() {
+//       _selectedIndex = index;
+//     });
 
-    if (_selectedIndex == 4) {
-      Navigator.push(
-        context,
-        MaterialPageRoute(
-          builder: (BuildContext context) {
-            return const UpdateProfile();
-          },
-        ),
-      );
-    }
-  }
+//     if (_selectedIndex == 1) {
+//       _buildSearchBar();
+//     }
 
-  Widget _buildSearchBar() {
-    return Padding(
-      padding: const EdgeInsets.all(8.0),
-      child: TextField(
-        decoration: InputDecoration(
-          hintText: 'Search...',
-          prefixIcon: const Icon(Icons.search),
-          border: OutlineInputBorder(
-            borderRadius: BorderRadius.circular(16.0),
-          ),
-          contentPadding: const EdgeInsets.symmetric(vertical: 8.0),
-        ),
-      ),
-    );
-  }
+//     if (_selectedIndex == 4) {
+//       Navigator.push(
+//         context,
+//         MaterialPageRoute(
+//           builder: (BuildContext context) {
+//             return const UpdateProfile();
+//           },
+//         ),
+//       );
+//     }
+//   }
 
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
+//   Widget _buildSearchBar() {
+//     return Padding(
+//       padding: const EdgeInsets.all(8.0),
+//       child: TextField(
+//         decoration: InputDecoration(
+//           hintText: 'Search...',
+//           prefixIcon: const Icon(Icons.search),
+//           border: OutlineInputBorder(
+//             borderRadius: BorderRadius.circular(16.0),
+//           ),
+//           contentPadding: const EdgeInsets.symmetric(vertical: 8.0),
+//         ),
+//       ),
+//     );
+//   }
+
+//   @override
+//   Widget build(BuildContext context) {
+//     return Scaffold(
       
-      body: Container(
-        color: Colors.green,
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            const SizedBox(height: 20),
-            Row(
-              children: [
-                const Text(
-                  'Profile',
-                  style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
-                ),
-                const Spacer(),
-                Padding(
-                  padding: const EdgeInsets.symmetric(horizontal: 16.0),
-                  child: Container(
-                    decoration: BoxDecoration(
-                      color: Colors.white,
-                      border: Border.all(
-                        color: Colors.white,
-                        width: 2.0,
-                      ),
-                      borderRadius:
-                          const BorderRadius.all(Radius.circular(10.0)),
-                    ),
-                    child: const Padding(
-                      padding: EdgeInsets.all(8.0),
-                      child: Text(
-                        'Task-seekers',
-                        style: TextStyle(
-                          fontSize: 16,
-                          fontWeight: FontWeight.bold,
-                          color: Colors.black,
-                        ),
-                      ),
-                    ),
-                  ),
-                ),
-              ],
-            ),
-            const SizedBox(height: 20),
-            const Flexible(
-              child: MyCustomForm(),
-            ),
-            const SizedBox(height: 20),
-            Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                const SizedBox(height: 10),
-                ElevatedButton(
-                  onPressed: () {
-                    Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                        builder: (BuildContext context) {
-                          return const UpdateProfile();
-                        },
-                      ),
-                    );
-                  },
-                  style: ElevatedButton.styleFrom(
-                    foregroundColor: Colors.black,
-                    backgroundColor: Colors.white,
-                  ),
-                  child: const Text('Update Profile'),
-                ),
-                const SizedBox(height: 10),
-                ElevatedButton(
-                  onPressed: () {
-                    Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                        builder: (BuildContext context) {
-                          return const ChangePassword();
-                        },
-                      ),
-                    );
-                  },
-                  style: ElevatedButton.styleFrom(
-                    foregroundColor: Colors.black,
-                    backgroundColor: Colors.white,
-                  ),
-                  child: const Text('Change Password'),
-                ),
-                const SizedBox(height: 10),
-              ],
-            ),
-          ],
-        ),
-      ),
+//       body: Container(
+//         color: Colors.green,
+//         child: Column(
+//           mainAxisAlignment: MainAxisAlignment.center,
+//           children: [
+//             const SizedBox(height: 20),
+//             Row(
+//               children: [
+//                 const Text(
+//                   'Profile',
+//                   style: TextStyle(fontSize: 24, fontWeight: FontWeight.bold),
+//                 ),
+//                 const Spacer(),
+//                 Padding(
+//                   padding: const EdgeInsets.symmetric(horizontal: 16.0),
+//                   child: Container(
+//                     decoration: BoxDecoration(
+//                       color: Colors.white,
+//                       border: Border.all(
+//                         color: Colors.white,
+//                         width: 2.0,
+//                       ),
+//                       borderRadius:
+//                           const BorderRadius.all(Radius.circular(10.0)),
+//                     ),
+//                     child: const Padding(
+//                       padding: EdgeInsets.all(8.0),
+//                       child: Text(
+//                         'Task-seekers',
+//                         style: TextStyle(
+//                           fontSize: 16,
+//                           fontWeight: FontWeight.bold,
+//                           color: Colors.black,
+//                         ),
+//                       ),
+//                     ),
+//                   ),
+//                 ),
+//               ],
+//             ),
+//             const SizedBox(height: 20),
+//             const Flexible(
+//               child: MyCustomForm(),
+//             ),
+//             const SizedBox(height: 20),
+//             Column(
+//               mainAxisAlignment: MainAxisAlignment.center,
+//               children: [
+//                 const SizedBox(height: 10),
+//                 ElevatedButton(
+//                   onPressed: () {
+//                     Navigator.push(
+//                       context,
+//                       MaterialPageRoute(
+//                         builder: (BuildContext context) {
+//                           return const UpdateProfile();
+//                         },
+//                       ),
+//                     );
+//                   },
+//                   style: ElevatedButton.styleFrom(
+//                     foregroundColor: Colors.black,
+//                     backgroundColor: Colors.white,
+//                   ),
+//                   child: const Text('Update Profile'),
+//                 ),
+//                 const SizedBox(height: 10),
+//                 ElevatedButton(
+//                   onPressed: () {
+//                     Navigator.push(
+//                       context,
+//                       MaterialPageRoute(
+//                         builder: (BuildContext context) {
+//                           return const ChangePassword();
+//                         },
+//                       ),
+//                     );
+//                   },
+//                   style: ElevatedButton.styleFrom(
+//                     foregroundColor: Colors.black,
+//                     backgroundColor: Colors.white,
+//                   ),
+//                   child: const Text('Change Password'),
+//                 ),
+//                 const SizedBox(height: 10),
+//               ],
+//             ),
+//           ],
+//         ),
+//       ),
 
-    );
-  }
-}
+//     );
+//   }
+// }
 
-class MyCustomForm extends StatefulWidget {
-  const MyCustomForm({Key? key}) : super(key: key);
+// class MyCustomForm extends StatefulWidget {
+//   const MyCustomForm({Key? key}) : super(key: key);
 
-  @override
-  MyCustomFormState createState() {
-    return MyCustomFormState();
-  }
-}
+//   @override
+//   MyCustomFormState createState() {
+//     return MyCustomFormState();
+//   }
+// }
 
-class MyCustomFormState extends State<MyCustomForm> {
-  final _formKey = GlobalKey<FormState>();
+// class MyCustomFormState extends State<MyCustomForm> {
+//   final _formKey = GlobalKey<FormState>();
 
-  @override
-  Widget build(BuildContext context) {
-    return Form(
-      key: _formKey,
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: <Widget>[
-          TextFormField(
-            decoration: const InputDecoration(
-              icon: Icon(Icons.android, color: Colors.white),
-              hintText: 'Enter your user ID',
-              labelText: 'UserID',
-              labelStyle: TextStyle(color: Colors.white),
-              enabledBorder: UnderlineInputBorder(
-                borderSide: BorderSide(color: Colors.white),
-              ),
-              focusedBorder: UnderlineInputBorder(
-                borderSide: BorderSide(color: Colors.white),
-              ),
-            ),
-          ),
-          TextFormField(
-            decoration: const InputDecoration(
-              icon: Icon(Icons.person, color: Colors.white),
-              hintText: 'Enter your name',
-              labelText: 'Name',
-              labelStyle: TextStyle(color: Colors.white),
-              enabledBorder: UnderlineInputBorder(
-                borderSide: BorderSide(color: Colors.white),
-              ),
-              focusedBorder: UnderlineInputBorder(
-                borderSide: BorderSide(color: Colors.white),
-              ),
-            ),
-          ),
-          TextFormField(
-            decoration: const InputDecoration(
-              icon: Icon(Icons.email, color: Colors.white),
-              hintText: 'Enter your email address',
-              labelText: 'Email Address',
-              labelStyle: TextStyle(color: Colors.white),
-              enabledBorder: UnderlineInputBorder(
-                borderSide: BorderSide(color: Colors.white),
-              ),
-              focusedBorder: UnderlineInputBorder(
-                borderSide: BorderSide(color: Colors.white),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
+//   @override
+//   Widget build(BuildContext context) {
+//     return Form(
+//       key: _formKey,
+//       child: Column(
+//         crossAxisAlignment: CrossAxisAlignment.start,
+//         children: <Widget>[
+//           TextFormField(
+//             decoration: const InputDecoration(
+//               icon: Icon(Icons.android, color: Colors.white),
+//               hintText: 'Enter your user ID',
+//               labelText: 'UserID',
+//               labelStyle: TextStyle(color: Colors.white),
+//               enabledBorder: UnderlineInputBorder(
+//                 borderSide: BorderSide(color: Colors.white),
+//               ),
+//               focusedBorder: UnderlineInputBorder(
+//                 borderSide: BorderSide(color: Colors.white),
+//               ),
+//             ),
+//           ),
+//           TextFormField(
+//             decoration: const InputDecoration(
+//               icon: Icon(Icons.person, color: Colors.white),
+//               hintText: 'Enter your name',
+//               labelText: 'Name',
+//               labelStyle: TextStyle(color: Colors.white),
+//               enabledBorder: UnderlineInputBorder(
+//                 borderSide: BorderSide(color: Colors.white),
+//               ),
+//               focusedBorder: UnderlineInputBorder(
+//                 borderSide: BorderSide(color: Colors.white),
+//               ),
+//             ),
+//           ),
+//           TextFormField(
+//             decoration: const InputDecoration(
+//               icon: Icon(Icons.email, color: Colors.white),
+//               hintText: 'Enter your email address',
+//               labelText: 'Email Address',
+//               labelStyle: TextStyle(color: Colors.white),
+//               enabledBorder: UnderlineInputBorder(
+//                 borderSide: BorderSide(color: Colors.white),
+//               ),
+//               focusedBorder: UnderlineInputBorder(
+//                 borderSide: BorderSide(color: Colors.white),
+//               ),
+//             ),
+//           ),
+//         ],
+//       ),
+//     );
+//   }
+// }
 
-class ChangePassword extends StatelessWidget {
-  const ChangePassword({Key? key}) : super(key: key);
+// class ChangePassword extends StatelessWidget {
+//   const ChangePassword({Key? key}) : super(key: key);
 
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Change Password'),
-      ),
-      body: const Center(
-        child: Text(
-          'This is the change password screen.',
-          style: TextStyle(color: Colors.black),
-        ),
-      ),
-    );
-  }
-}
+//   @override
+//   Widget build(BuildContext context) {
+//     return Scaffold(
+//       appBar: AppBar(
+//         title: const Text('Change Password'),
+//       ),
+//       body: const Center(
+//         child: Text(
+//           'This is the change password screen.',
+//           style: TextStyle(color: Colors.black),
+//         ),
+//       ),
+//     );
+//   }
+// }
